Commands - just implement commandexecutor + lots of cleanup EconomyHandler - get rid of useless interface Permission - refactor PodcrashPlugin - refactor + we think we know the source of mem leak
New code conventions
Everything else is basic code cleanup done automatically by IDE

+++ MCEngine/src/main/java/com/podcrash/api/mc/game/TeamEnum.java
@@ -35,7 +35,7 @@ public enum TeamEnum {
         this.name = name;
         this.data = data;
         this.byteData = byteData;
-        this.intData = (int) byteData;
+        this.intData = byteData;
         this.chatColor = chatColor;
         this.scoreboardColor = scoreboardColor;
         this.dyeColor = dyeColor;
